Separate basics market data from calc
Use separate testing identifiers for different projects
