+++ modules/calc/src/test/java/com/opengamma/strata/calc/marketdata/BuiltScenarioMarketDataTest.java
@@ -11,7 +11,6 @@ import static com.opengamma.strata.collect.TestHelper.date;
 import org.testng.annotations.Test;
 
 import com.opengamma.strata.basics.market.MarketDataBox;
-import com.opengamma.strata.basics.market.TestObservableId;
 import com.opengamma.strata.collect.result.FailureException;
 import com.opengamma.strata.collect.result.FailureReason;
 import com.opengamma.strata.collect.result.Result;
