+++ modules/calc/src/test/java/com/opengamma/strata/calc/marketdata/TestObservableId.java
@@ -16,25 +16,43 @@ import com.opengamma.strata.basics.market.ObservableSource;
 /**
  * An observable ID implementation used in tests.
  */
-public class TestObservableId implements ObservableId, Serializable {
+public class TestObservableId
+    implements ObservableId, Serializable {
 
   private static final long serialVersionUID = 1L;
 
-  private final String id;
+  private final StandardId id;
+
   private final ObservableSource observableSource;
 
   public static TestObservableId of(String id) {
     return new TestObservableId(id, ObservableSource.NONE);
   }
 
-  public TestObservableId(String id, ObservableSource obsSource) {
-    this.id = id;
+  public static TestObservableId of(String id, ObservableSource obsSource) {
+    return new TestObservableId(id, obsSource);
+  }
+
+  public static TestObservableId of(StandardId id) {
+    return new TestObservableId(id, ObservableSource.NONE);
+  }
+
+  public static TestObservableId of(StandardId id, ObservableSource obsSource) {
+    return new TestObservableId(id, obsSource);
+  }
+
+  TestObservableId(String id, ObservableSource obsSource) {
+    this(StandardId.of("test", id), obsSource);
+  }
+
+  TestObservableId(StandardId id, ObservableSource obsSource) {
     this.observableSource = obsSource;
+    this.id = id;
   }
 
   @Override
   public StandardId getStandardId() {
-    return StandardId.of("test", id);
+    return id;
   }
 
   @Override
@@ -52,21 +70,26 @@ public class TestObservableId implements ObservableId, Serializable {
     return new TestObservableId(id, obsSource);
   }
 
+  //-------------------------------------------------------------------------
   @Override
-  public boolean equals(Object o) {
-    if (this == o) {
+  public boolean equals(Object obj) {
+    if (this == obj) {
       return true;
     }
-    if (o == null || getClass() != o.getClass()) {
+    if (obj == null || getClass() != obj.getClass()) {
       return false;
     }
-    TestObservableId that = (TestObservableId) o;
-    return Objects.equals(id, that.id) &&
-        Objects.equals(observableSource, that.observableSource);
+    TestObservableId that = (TestObservableId) obj;
+    return Objects.equals(id, that.id);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(id, observableSource);
+    return Objects.hash(id);
+  }
+
+  @Override
+  public String toString() {
+    return "TestObservableId [id=" + id + ", observableSource=" + observableSource + "]";
   }
 }
