+++ modules/calc/src/test/java/com/opengamma/strata/calc/runner/DefaultCalculationTaskRunnerTest.java
@@ -24,15 +24,15 @@ import com.google.common.util.concurrent.MoreExecutors;
 import com.opengamma.strata.basics.CalculationTarget;
 import com.opengamma.strata.basics.currency.Currency;
 import com.opengamma.strata.basics.market.ReferenceData;
-import com.opengamma.strata.basics.market.TestObservableId;
-import com.opengamma.strata.calc.ScenarioMarketData;
 import com.opengamma.strata.calc.Column;
 import com.opengamma.strata.calc.ImmutableScenarioMarketData;
 import com.opengamma.strata.calc.Measure;
 import com.opengamma.strata.calc.Measures;
 import com.opengamma.strata.calc.Results;
+import com.opengamma.strata.calc.ScenarioMarketData;
 import com.opengamma.strata.calc.marketdata.FunctionRequirements;
 import com.opengamma.strata.calc.marketdata.TestId;
+import com.opengamma.strata.calc.marketdata.TestObservableId;
 import com.opengamma.strata.calc.result.ScenarioResult;
 import com.opengamma.strata.calc.runner.CalculationTaskTest.TestTarget;
 import com.opengamma.strata.collect.result.Result;
